requirements-tests.txt
-r requirements.txt
mock==2.0.0
pytest==3.0.7
pytest-bdd==2.18.2
pytest-django==3.1.2
pytest-cov==2.5.1
pytest-runner==2.11.1
requests==2.13.0
